Move countdown animation to separate screen
